+++ tests/test_mps.py
@@ -210,7 +210,7 @@ def test_compute_K():
     pairs = [(0, 1), (2, 3), (4, 5)]  # singlets on a 3x2 grid -> k_y = pi
     psi = mps.MPS.from_singlets(spin_half, 6, pairs, bc='infinite')
     psi.test_sanity()
-    lat = Square(3, 2, spin_half, order='default', bc_MPS='infinite')
+    lat = Square(3, 2, spin_half, order='default', bc_MPS='infinite', bc='periodic')
     U, W, q, ov, te = psi.compute_K(lat, verbose=100)
     assert (ov == -1.)
     npt.assert_array_equal(W, [1.])